requirements.txt
absl-py==2.1.0
aiohappyeyeballs==2.3.5
aiohttp==3.10.3
aiosignal==1.3.1
antlr4-python3-runtime==4.9.3
async-timeout==4.0.3
attrs==24.2.0
beartype==0.18.5
cachetools==5.4.0
certifi==2024.7.4
charset-normalizer==3.3.2
clip==0.2.0
contourpy==1.2.1
cycler==0.12.1
diffusers==0.18.2
ema-pytorch==0.4.3
et-xmlfile==1.1.0
filelock==3.15.4
fonttools==4.53.1
frozenlist==1.4.1
fsspec==2024.6.1
google-auth==2.33.0
google-auth-oauthlib==1.0.0
grpcio==1.65.4
huggingface-hub==0.24.5
idna==3.7
imageio==2.35.0
importlib_metadata==8.2.0
importlib_resources==6.4.2
Jinja2==3.1.4
joblib==1.4.2
kiwisolver==1.4.5
lazy_loader==0.4
lightning-utilities==0.11.6
lpips==0.1.4
Markdown==3.6
MarkupSafe==2.1.5
matplotlib==3.9.2
mpmath==1.3.0
multidict==6.0.5
natsort==8.4.0
networkx==3.2.1
numpy==1.24.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.19.3
nvidia-nvjitlink-cu12==12.6.20
nvidia-nvtx-cu12==12.1.105
oauthlib==3.2.2
omegaconf==2.3.0
opencv-python==4.8.0.74
openpyxl==3.1.2
orderedmultidict==1.0.1
packaging==24.1
pandas==2.0.3
pillow==10.4.0
protobuf==5.27.3
pyasn1==0.6.0
pyasn1_modules==0.4.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
pytorch-lightning==2.0.5
pytz==2024.1
PyWavelets==1.6.0
PyYAML==6.0.2
regex==2024.7.24
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9
safetensors==0.4.4
scikit-image==0.21.0
scikit-learn==1.3.0
scikit-misc==0.2.0
scipy==1.10.1
seaborn==0.13.0
sentry-sdk==1.28.1
setproctitle==1.3.2
simplejson==3.19.2
six==1.16.0
sympy==1.13.2
taming-transformers==0.0.1
tb-nightly==2.14.0a20230714
tensorboard-data-server==0.7.1
tensorboardX==2.6.1
threadpoolctl==3.5.0
tifffile==2024.8.10
timm==0.9.16
tokenizers==0.13.3
tomli==2.0.1
torch==2.2.0
torchinfo==1.8.0
torchmetrics==1.4.1
torchvision==0.17.0
tqdm==4.66.5
traitlets==5.14.2
transformers==4.30.2
triton==2.2.0
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
Werkzeug==3.0.3
yarl==1.9.4
zipp==3.20.0
